refactor resource edit page
