+++ src/components/resource/edit-custom-title/resource-edit-custom-title.js
@@ -1,15 +1,13 @@
 import React, { Component } from 'react';
-import PropTypes from 'prop-types';
 import {
   Form,
-  FormSpy,
 } from 'react-final-form';
 import arrayMutators from 'final-form-arrays';
 import createFocusDecorator from 'final-form-focus';
 import { FormattedMessage } from 'react-intl';
 
-import update from 'lodash/fp/update';
 import hasIn from 'lodash/hasIn';
+import update from 'lodash/fp/update';
 
 import { withStripes } from '@folio/stripes-core';
 import {
@@ -19,7 +17,6 @@ import {
   Icon,
   Modal,
   ModalFooter,
-  PaneFooter,
 } from '@folio/stripes/components';
 
 import CoverageStatementFields, { coverageStatementDecorator } from '../_fields/coverage-statement';
@@ -27,7 +24,6 @@ import CustomEmbargoFields, { getEmbargoInitial } from '../_fields/custom-embarg
 import CustomUrlFields from '../_fields/custom-url';
 import CoverageFields from '../_fields/resource-coverage-fields';
 import VisibilityField from '../_fields/visibility';
-import CoverageDateList from '../../coverage-date-list';
 import { CustomLabelsEditSection } from '../../custom-labels-section';
 import DetailsView from '../../details-view';
 import NavigationModal from '../../navigation-modal';
@@ -35,9 +31,10 @@ import ProxySelectField from '../../proxy-select';
 import AccessTypeEditSection from '../../access-type-edit-section';
 import Toaster from '../../toaster';
 
+import resourceEditProptypes from '../recource-edit-proptypes';
+
 import {
   processErrors,
-  isBookPublicationType,
   getUserDefinedFields,
   getAccessTypeId,
   getProxyTypesRecords,
@@ -49,7 +46,6 @@ import { CustomLabelsAccordion } from '../../../features';
 import {
   historyActions,
   coverageStatementExistenceStatuses,
-  accessTypesReduxStateShape,
 } from '../../../constants';
 
 
@@ -91,25 +87,13 @@ class ResourceEditCustomTitle extends Component {
     return proxyTypes.request.isResolved;
   }
 
-  static propTypes = {
-    accessStatusTypes: accessTypesReduxStateShape.isRequired,
-    model: PropTypes.object.isRequired,
-    onCancel: PropTypes.func.isRequired,
-    onSubmit: PropTypes.func.isRequired,
-    proxyTypes: PropTypes.object.isRequired,
-    stripes: PropTypes.shape({
-      hasPerm: PropTypes.func.isRequired,
-    }),
-  };
+  static propTypes = resourceEditProptypes;
 
   constructor(props) {
     super(props);
     this.state = {
       resourceSelected: this.props.model.isSelected,
-      showSelectionModal: false,
-      allowFormToSubmit: false,
       wasProxyTypesLoaded: this.props.proxyTypes.request.isResolved,
-      formValues: {},
       initialValues: ResourceEditCustomTitle.getInitialValuesFromModel(this.props.model, this.props.proxyTypes),
       sections: {
         resourceShowHoldingStatus: true,
@@ -126,10 +110,6 @@ class ResourceEditCustomTitle extends Component {
 
     const { proxyTypes } = nextProps;
     const isProxyTypesLoaded = ResourceEditCustomTitle.isProxyTypesLoaded(proxyTypes);
-
-    if (nextProps.model.destroy.errors.length) {
-      stateUpdates.showSelectionModal = false;
-    }
 
     if (nextProps.model.isSelected !== prevState.initialValues.isSelected) {
       Object.assign(stateUpdates, {
@@ -159,22 +139,12 @@ class ResourceEditCustomTitle extends Component {
     this.setState({ sections });
   };
 
-  getSectionHeader = (translationKey) => {
-    return (
-      <Headline
-        size="large"
-        tag="h3"
-      >
-        <FormattedMessage id={translationKey} />
-      </Headline>);
-  };
-
   handleRemoveResourceFromHoldings = () => {
     this.setState({
       formValues: {
         isSelected: false
       }
-    }, () => { this.handleOnSubmit(this.state.formValues); });
+    }, () => { this.props.handleOnSubmit(this.state.formValues); });
   };
 
   handleSelectionToggle = (e) => {
@@ -183,65 +153,14 @@ class ResourceEditCustomTitle extends Component {
     });
   };
 
-  commitSelectionToggle = () => {
-    this.setState({
-      allowFormToSubmit: true
-    }, () => { this.handleOnSubmit(this.state.formValues); });
-  };
-
   cancelSelectionToggle = (change) => {
+    this.props.closeSelectionModal();
+
     this.setState({
-      showSelectionModal: false,
       resourceSelected: true,
     }, () => {
       change('isSelected', true);
     });
-  };
-
-  handleOnSubmit = (values) => {
-    if (this.state.allowFormToSubmit === false && values.isSelected === false) {
-      this.setState({
-        showSelectionModal: true,
-        formValues: values
-      });
-    } else {
-      this.setState({
-        allowFormToSubmit: false,
-        formValues: {}
-      }, () => {
-        this.props.onSubmit(values);
-      });
-    }
-  };
-
-  renderCoverageDates = () => {
-    return (
-      <FormSpy subscription={{ values: true }}>
-        {({ values }) => {
-          const { model } = this.props;
-          const { customCoverages: customCoverageDateValues } = values;
-          let coverageDates = model.managedCoverages;
-
-          if (customCoverageDateValues && customCoverageDateValues.length > 0) {
-            coverageDates = customCoverageDateValues;
-          }
-
-          const nonEmptyCoverageDates = coverageDates
-            .filter((currentCoverageDate) => Object.keys(currentCoverageDate).length !== 0);
-
-          if (nonEmptyCoverageDates.length === 0) {
-            return null;
-          }
-
-          return (
-            <CoverageDateList
-              coverageArray={nonEmptyCoverageDates}
-              isYearOnly={isBookPublicationType(model.publicationType)}
-            />
-          );
-        }}
-      </FormSpy>
-    );
   };
 
   getActionMenu = () => {
@@ -266,51 +185,21 @@ class ResourceEditCustomTitle extends Component {
     );
   }
 
-  getFooter = (pristine, reset) => {
-    const { model } = this.props;
-
-    const cancelButton = (
-      <Button
-        data-test-eholdings-provider-edit-cancel-button
-        buttonStyle="default mega"
-        disabled={model.update.isPending || model.destroy.isPending || pristine}
-        onClick={reset}
-        marginBottom0
-      >
-        <FormattedMessage id="stripes-components.cancel" />
-      </Button>
-    );
-
-    const saveButton = (
-      <Button
-        buttonStyle="primary mega"
-        data-test-eholdings-resource-save-button
-        disabled={pristine || model.update.isPending || model.destroy.isPending}
-        marginBottom0
-        type="submit"
-      >
-        <FormattedMessage id="stripes-components.saveAndClose" />
-      </Button>
-    );
-
-    return (
-      <PaneFooter
-        renderStart={cancelButton}
-        renderEnd={saveButton}
-      />
-    );
-  }
-
   render() {
     const {
       model,
       proxyTypes,
       onCancel,
       accessStatusTypes,
+      showSelectionModal,
+      handelDeleteConfirmation,
+      handleOnSubmit,
+      getFooter,
+      getSectionHeader,
+      renderCoverageDates,
     } = this.props;
 
     const {
-      showSelectionModal,
       resourceSelected,
       sections,
       initialValues,
@@ -325,7 +214,7 @@ class ResourceEditCustomTitle extends Component {
 
     return (
       <Form
-        onSubmit={this.handleOnSubmit}
+        onSubmit={handleOnSubmit}
         decorators={[coverageStatementDecorator, focusOnErrors]}
         mutators={{ ...arrayMutators }}
         initialValues={initialValues}
@@ -341,11 +230,11 @@ class ResourceEditCustomTitle extends Component {
                 actionMenu={this.getActionMenu()}
                 handleExpandAll={this.toggleAllSections}
                 sections={sections}
-                footer={this.getFooter(pristine, reset)}
+                footer={getFooter(pristine, reset)}
                 bodyContent={(
                   <>
                     <Accordion
-                      label={this.getSectionHeader('ui-eholdings.label.holdingStatus')}
+                      label={getSectionHeader('ui-eholdings.label.holdingStatus')}
                       open={sections.resourceShowHoldingStatus}
                       id="resourceShowHoldingStatus"
                       onToggle={this.toggleSection}
@@ -373,7 +262,7 @@ class ResourceEditCustomTitle extends Component {
                     />
 
                     <Accordion
-                      label={this.getSectionHeader('ui-eholdings.resource.resourceSettings')}
+                      label={getSectionHeader('ui-eholdings.resource.resourceSettings')}
                       open={sections.resourceShowSettings}
                       id="resourceShowSettings"
                       onToggle={this.toggleSection}
@@ -403,7 +292,7 @@ class ResourceEditCustomTitle extends Component {
                     </Accordion>
 
                     <Accordion
-                      label={this.getSectionHeader('ui-eholdings.label.coverageSettings')}
+                      label={getSectionHeader('ui-eholdings.label.coverageSettings')}
                       open={sections.resourceShowCoverageSettings}
                       id="resourceShowCoverageSettings"
                       onToggle={this.toggleSection}
@@ -417,7 +306,7 @@ class ResourceEditCustomTitle extends Component {
                             <FormattedMessage id="ui-eholdings.label.coverageStatement" />
                           </Headline>
                           <CoverageStatementFields
-                            coverageDates={this.renderCoverageDates()}
+                            coverageDates={renderCoverageDates()}
                           />
 
                           <Headline tag="h4">
@@ -453,7 +342,7 @@ class ResourceEditCustomTitle extends Component {
                     data-test-eholdings-resource-deselection-confirmation-modal-yes
                     buttonStyle="primary"
                     disabled={model.destroy.isPending}
-                    onClick={this.commitSelectionToggle}
+                    onClick={handelDeleteConfirmation}
                   >
                     {(model.destroy.isPending ?
                       <FormattedMessage id="ui-eholdings.resource.modal.buttonWorking" /> :
